// src/app/webinar/types/webinar-type.ts
export interface Webinar {
  organization: string;
  sponsoredBy: string;
  followers: number;
  timeAgo: string;
  orgLogo: string;
  bannerUrl: string;
  credits: string;
  expiresIn: string;
  title: string;
  liveDate: string;
  keyHighlight: string;
  likes: number;
  comments: number;
  watching: number;
}